requirements.txt
streamlit
requests
nltk
PyPDF2
python-docx
googlesearch-python